requirements.txt
python-dotenv~=1.0.0
requests~=2.30.0
motor~=3.1.2
pymongo~=4.3.3
aiohttp~=3.8.4
discord.py